requirements.txt
numpy
scipy
tensorflow-gpu
Pillow
matplotlib
jupyter
tqdm